+++ changelog.md
@@ -1,4 +1,2 @@
 ### Packed Up 1.0.27
-- Updated to core library 1.1
-- Backpack inventories will be saved on more occasions
-- Added advancements to unlock the backpack recipes
+- Initial release of Packed Up for Minecraft 1.19.3